requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
Flask-Login==0.6.3
PyPDF2==3.0.1
pycryptodome==3.23.0
requests==2.31.0
Werkzeug==3.0.1